==> analyses/07-spatial_prediction/01-Pv_areal_CAR_MCMC.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
# Bayesian Hierarchial Spatial Province Models for _P. vivax_ among Asymptomatic Adults in the DRC"


```{r, include=FALSE}
knitr::opts_chunk$set(echo = T, warning = F, message = F, eval = T, results = 'hide',
                      fig.align = 'center', fig.width = 8, fig.height = 8)

```

```{r}
source("~/Documents/GitHub/VivID_Epi/R/00-functions_basic.R") 
source("~/Documents/GitHub/VivID_Epi/R/00-functions_epi.R") 
source("~/Documents/GitHub/VivID_Epi/R/00-MCMC_diagnostics.R")
library(tidyverse)
library(srvyr) #wrap the survey package in dplyr syntax
library(CARBayes)
library(HDInterval)
library(raster)
set.seed(48)
tol <- 1e-3

#......................
# Import Data
#......................
dt <- readRDS("~/Documents/GitHub/VivID_Epi/data/derived_data/vividepi_recode_completecases.rds")
dtsrvy <- makecd2013survey(survey = dt)
mp <- readRDS("~/Documents/GitHub/VivID_Epi/data/derived_data/basic_cluster_mapping_data.rds")
ge <- readRDS(file = "~/Documents/GitHub/VivID_Epi/data/raw_data/dhsdata/VivIDge.RDS")

```

```{r}

#------------------------------------------------------------------------
# Subset to Pv
#------------------------------------------------------------------------
pvprov.weighted <- mp$data[mp$plsmdmspec == "pv18s" & mp$maplvl == "adm1name"][[1]]
# vectors have destroyed spatial class, need to remake
pvprov.weighted <- sf::st_as_sf(pvprov.weighted)
sf::st_crs(pvprov.weighted) <-  sf::st_crs(ge)
# need ints (binomail prob), so will round
pvprov.weighted <- pvprov.weighted %>% 
  dplyr::mutate(plsmdn = round(plsmdn, 0),
                n = round(n, 0))

pvprov.weighted.nosf <- pvprov.weighted
sf::st_geometry(pvprov.weighted.nosf) <- NULL

```

```{r}
#..............................................................
# Import the Covariates
#..............................................................
pvcovar <- readRDS("~/Documents/GitHub/VivID_Epi/data/derived_data/vividepi_prov_covars_bayesian_fit.RDS")
# combine
pvprov.weighted.nosf <- dplyr::left_join(pvprov.weighted.nosf, pvcovar, by = "adm1name")


```


## Hierarchial Bayesian Spatial Models for Province Prevalence 
### Make Model Framework
NB. When $\rho$ is `NULL`, it is estimated within the CAR model. based on the _Leroux et al. 2000_ specifiation of the Gaussian Markov Random Field.

```{r, results='asis'}

#-------------------------------------------------------------------------
# Conditional Autoregressive Spatial Model 
#-------------------------------------------------------------------------
#......................
# Make Adjacency Matrix for Pv 
#......................
W.nb <- spdep::poly2nb(sf::as_Spatial(pvprov.weighted), row.names = pvprov.weighted$adm1name)
W <- spdep::nb2mat(W.nb, style = "B") # binary weights taking values zero or one (only one is recorded)

#......................
# Make Model Framework
#......................
prov.covar.names <- c("precip_scale", "crop_scale", "nightlight_scale")
mod.framework <- tibble(name = c("CAR_intercept", "CAR_covar"),
                        formula = c("plsmdn ~ 1", 
                                    paste0("plsmdn ~ ", paste(prov.covar.names, collapse = " + "))),
                        burnin = 1e3,
                        n.sample = 1e4 + 1e3,
                        family = "binomial"
                        )


mod.framework$trials <- lapply(1:nrow(mod.framework), function(x) return(pvprov.weighted.nosf$n))
mod.framework$data <- lapply(1:nrow(mod.framework), function(x) return(pvprov.weighted.nosf))
mod.framework$W <- lapply(1:nrow(mod.framework), function(x) return(W))

mod.framework %>% 
  dplyr::select(-c("data", "trials", "W")) %>% 
  knitr::kable(.)

```


#### Diagnostic Chains
Run out four diagnostic chains to check convergence. 
```{r}

# replicate this four times for our four chains
mod.framework <- lapply(1:4, function(x) return(mod.framework)) %>% 
  dplyr::bind_rows() %>% 
  dplyr::arrange(name)

#......................
# Make a wrapper for CARBAYES
#......................
wrap_S.CARleroux <- function(name, formula, family, trials, W, data, burnin, n.sample){
  
  formvec <- paste(formula, collapse = "")
  betacount <- stringr::str_count(formvec, "\\+") + 2 # need intercept and betas
  betacount <- ifelse(grepl("1", formvec), 1, betacount) # corner case of just intercept
  prior.var.betavec <- rep(5e4, betacount) # note prior setting here
  
  # rho here is NULL by default and esimtated in the model
  ret <- CARBayes::S.CARleroux(formula = as.formula(formula), 
                               family = family, 
                               trials = trials, 
                               W = W,
                               data = data,
                               burnin = burnin, 
                               prior.var.beta = prior.var.betavec,
                               n.sample = n.sample)
  
  
  return(ret)
}


mod.framework$MCMC <- purrr::pmap(mod.framework, wrap_S.CARleroux)

```

#### MCMC Diagnostics 
Note, our goal is to have the effective sample sizes (`n.effective`) to be at least 500 for all parameters. In addition, we will visually check for convergence among the four chains. 

##### Param Diagnostics
```{r}
#-------------------------------------------------------------------------
# MCMC Diagnostics
#-------------------------------------------------------------------------
mod.framework$mcmc.modsum <- purrr::map(mod.framework$MCMC, print) # note, print is overloaded here
# CI mean, mode, median
mod.framework$summresults <- purrr::map(mod.framework$mcmc.modsum, "summary.results")
# get samples for HPD
mod.framework$samples <- purrr::map(mod.framework$MCMC, "samples")
# get HPD
mod.framework$HPD <- purrr::map(mod.framework$samples, HDInterval::hdi, credMass = 0.95)
# subset HPD to relevant params
mod.framework$HPD.summ <- purrr::map(mod.framework$HPD, 
                                     function(x){
                                       ret <- x[ names(x) %in% c("beta", "rho", "tau2") ]
                                       ret <- lapply(ret, function(y){
                                         ret <- t(y)
                                         colnames(ret) <- c("HPD 0.025", "HPD 0.975")
                                         return(ret) }) %>% 
                                         do.call("rbind.data.frame", .)
                                     })

mod.framework$summresults <- purrr::pmap(mod.framework[,c("summresults", "HPD.summ")], 
                                         function(summresults, HPD.summ){
                                           pars <- rownames(summresults)
                                           ret <- cbind.data.frame(pars = pars,
                                                                   cbind( summresults, HPD.summ))
                                           return(ret)
                                         })

```

###### Summary Stats
```{r, results='asis'}
mod.framework %>% 
  dplyr::select(c("name", "formula", "summresults")) %>% 
  tidyr::unnest(cols = "summresults") %>% 
    DT::datatable(., 
                rownames = F,
                options = list(
                  searching = F,
                  pageLength = 5,
                  dom = 'Bfrtip', 
                  buttons = c('csv')))
```


##### Diagnostic Chains
Note, there are too many chains for visualizing here. Writing out to local directory. 
```{r}

#............................
# Let's Look at the Chains
#............................
mod.framework$betachains <- purrr::map(mod.framework$samples, "beta")
mod.framework$phichains <- purrr::map(mod.framework$samples, "phi")
mod.framework$tau2chains <- purrr::map(mod.framework$samples, "tau2")
mod.framework$rhochains <- purrr::map(mod.framework$samples, "rho")
mod.framework$fittedchains <- purrr::map(mod.framework$samples, "fitted")


chains <- mod.framework %>% 
  dplyr::select(c("name", "betachains", "phichains", "tau2chains", "rhochains", "fittedchains")) %>% 
  dplyr::group_by(name) %>% 
  tidyr::nest()

mydiagnosticdir <- "~/Documents/GitHub/VivID_Epi/analyses/07-spatial_prediction/prov_map_diagnostic_chains/"
dir.create(mydiagnosticdir, recursive = T)
wrap_chain_plotter.carbayes.diag(diag.dir = mydiagnosticdir, chains = chains)


```

## Final Long Chain
Now, we will run out one final long chain with 1e5 burnin iterations and 1e8 sampling iterations for each of our base models. 

```{r}

mod.framework.long <- mod.framework[,c("name", "formula", "family", "trials", "W", "data", "burnin", "n.sample")] %>% 
  dplyr::filter(!duplicated(.))

mod.framework.long$burnin <- 1e3
mod.framework.long$n.sample <- 1e5 + 1e3

mod.framework.long$MCMC <- purrr::pmap(mod.framework.long, wrap_S.CARleroux)


```

#### Long Chain MCMC Diagnostics 
Note, our goal is still to have the effective sample sizes (`n.effective`) to be at least 1,000 for all parameters. In addition, we will visually check for convergence among the four chains. 

##### Long Chain Param Diagnostics

```{r}
#-------------------------------------------------------------------------
# MCMC Diagnostics
#-------------------------------------------------------------------------
mod.framework.long$mcmc.modsum <- purrr::map(mod.framework.long$MCMC, print) # note, print is overloaded here
mod.framework.long$summresults <- purrr::map(mod.framework.long$mcmc.modsum, "summary.results")
# get samples for HPD
mod.framework.long$samples <- purrr::map(mod.framework.long$MCMC, "samples")
# get HPD
mod.framework.long$HPD <- purrr::map(mod.framework.long$samples, HDInterval::hdi, credMass = 0.95)
# subset HPD to relevant params
mod.framework.long$HPD.summ <- purrr::map(mod.framework.long$HPD, 
                                     function(x){
                                       ret <- x[ names(x) %in% c("beta", "rho", "tau2") ]
                                       ret <- lapply(ret, function(y){
                                         ret <- t(y)
                                         colnames(ret) <- c("HPD 0.025", "HPD 0.975")
                                         return(ret) }) %>% 
                                         do.call("rbind.data.frame", .)
                                     })

mod.framework.long$summresultsdf <- purrr::pmap(mod.framework.long[,c("summresults", "HPD.summ")], 
                                         function(summresults, HPD.summ){
                                           pars <- rownames(summresults)
                                           ret <- cbind.data.frame(pars = pars,
                                                                   cbind( summresults, HPD.summ))
                                           return(ret)
                                         })


#..............................................................
# Means for comparison to median
#..............................................................
mod.framework.long$posteriormeans <- purrr::pmap(mod.framework.long[,c("samples", "summresults")], 
                                            function(samples, summresults){
  betas <- apply(samples$beta, 2, mean)
  names(betas) <- row.names(summresults)[!row.names(summresults) %in% c("tau2", "rho")]
  tau2 <-  apply(samples$tau2, 2, mean)
  names(tau2) <- "tau2"
  rho <- apply(samples$rho, 2, mean)
  names(rho) <- "rho"
  
  params <- matrix(NA, nrow = 1, ncol = length(c(betas, tau2, rho)))
  params <- c(betas, tau2, rho)
  params <- data.frame(covar = names(params), postmean = params)
  return(params)
})

#..............................................................
# also want to get phis here for each province
#..............................................................
mod.framework.long$posteriorphis <- purrr::map(mod.framework.long$samples, 
                                            function(smpl){
                                              
  phimean <- apply(smpl$phi, 2, mean)
  phimedian <- apply(smpl$phi, 2, median)
  phiperc025 <- apply(smpl$phi, 2, quantile, 0.025)
  phiperc975 <- apply(smpl$phi, 2, quantile, 0.975)
  
  phihpd <- HDInterval::hdi(smpl$phi, credMass = 0.95)
  
  phineff <- apply(smpl$phi, 2, coda::effectiveSize)
  phis <- list(phimean, phimedian, phiperc025, phiperc975, phihpd, phineff)
  phis <- do.call("rbind.data.frame", phis)
  # bind together
  phis <- cbind.data.frame(statistic = c("Mean", "Median", "Perc0.025", "Perc97.5",
                                         "HPD 0.025", "HPD 0.975", "Neff"),
                   phis)
                   
  # put names in 
  colnames(phis)[2:ncol(phis)] <- pvprov.weighted$adm1name
  return(phis)
  
})


mod.framework.long.res <- mod.framework.long %>% 
  dplyr::select(c("name", "formula", "summresultsdf")) %>% 
  tidyr::unnest(cols = "summresultsdf") 
```

###### Summary Stats, Percentiles
```{r, results='asis'}
mod.framework.long.res %>% 
    DT::datatable(., 
                rownames = F,
                options = list(
                  searching = F,
                  pageLength = 20,
                  dom = 'Bfrtip', 
                  buttons = c('csv')))
```

###### Summary Stats, Means
```{r, results='asis'}
mod.framework.long %>% 
  dplyr::select(c("name", "formula", "posteriormeans")) %>% 
  tidyr::unnest(cols = "posteriormeans") %>% 
    DT::datatable(., 
                rownames = F,
                options = list(
                  searching = F,
                  pageLength = 20,
                  dom = 'Bfrtip', 
                  buttons = c('csv')))
```

###### Summary Stats, Phis
```{r, results='asis'}
mod.framework.long %>% 
  dplyr::select(c("name", "formula", "posteriorphis")) %>% 
  dplyr::mutate(var = "phi") %>% 
  tidyr::unnest(cols = "posteriorphis") %>% 
    DT::datatable(., 
                rownames = F,
                options = list(
                  searching = F,
                  pageLength = 20,
                  dom = 'Bfrtip', 
                  buttons = c('csv')))
```



Note, still a lot of chains for visualizing here. Writing out to local directory. 
```{r}

#............................
# Let's Look at the final chain
#............................
mod.framework.long$samples <- purrr::map(mod.framework.long$MCMC, "samples")
mod.framework.long$betachains <- purrr::map(mod.framework.long$samples, "beta")
mod.framework.long$phichains <- purrr::map(mod.framework.long$samples, "phi")
mod.framework.long$tau2chains <- purrr::map(mod.framework.long$samples, "tau2")
mod.framework.long$rhochains <- purrr::map(mod.framework.long$samples, "rho")
mod.framework.long$fittedchains <- purrr::map(mod.framework.long$samples, "fitted")


chains <- mod.framework.long %>% 
  dplyr::select(c("name", "betachains", "phichains", "tau2chains", "rhochains", "fittedchains")) %>% 
  dplyr::group_by(name) %>% 
  tidyr::nest()


myfinalchain <- "~/Documents/GitHub/VivID_Epi/analyses/07-spatial_prediction/prov_map_final_long_chain/"
dir.create(myfinalchain, recursive = T)
wrap_chain_plotter.carbayes.final(final.dir = myfinalchain, chains = chains)
```


### Long Chain DIC

```{r}


mod.framework.long$modfit <- purrr::map(mod.framework.long$mcmc.modsum, "modelfit")

# note in order to be consistent, with the PrevMap approach,
# I want Gelman's DIC: DICg = mu + sigma
# can get this from the log likelihood provided by CarBayes. 

# CarBayes nicely outputs the Fitted Values, from which I can calculate a log-likelihood
dimnames(mod.framework.long$W[[1]])[[1]]

mod.framework.long$LL <- purrr::pmap(mod.framework.long[, c("data", "trials", "MCMC")], 
  function(data, trials, MCMC){
  Yi <- data$plsmdn
  ni <- trials
  Ti <- MCMC$samples$fitted
  
  LL <- c()
  for(s in 1:nrow(Ti)){ # for every sim
    LL.iter <- 0
    for(i in 1:ncol(Ti)){ # for every spatial unit
      LL.iter <- dbinom(x = Yi[i], size = ni[i], prob = Ti[s, i]/ni[i], log = T) + LL.iter
      LL.iter
    }
    LL <- append(LL.iter, LL)
    
  }
  
  LL <- return(LL)
})


mod.framework.long$DICg <- purrr::map(mod.framework.long$LL, function(x){
  mu = mean(x)
  sigma = var(x)/2
  DICg = mu + sigma
  return(DICg)
})


mod.framework.long$DICs <- purrr::map(mod.framework.long$modfit, "DIC")


```

```{r, results='asis'}
data.frame(name = mod.framework.long$name, DICg = unlist(mod.framework.long$DICg),
           DICs = unlist(mod.framework.long$DICs)) %>% 
  DT::datatable(., 
                rownames = F,
                options = list(
                  searching = F,
                  pageLength = 6,
                  dom = 'Bfrtip', 
                  buttons = c('csv')))

```

```{r}

# save out this RDS file
saveRDS(mod.framework.long, 
        file = "~/Documents/GitHub/VivID_Epi/analyses/07-spatial_prediction/prov_map_final_long_chain/Province_Bayes_Hierarchial_Model_longchain.rds")

```
