==> analyses/07-spatial_prediction/02-Pv_clst_map_GaussianProc_MCMC_frontend.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
# Bayesian Hierarchial Spatial Cluster Models 

```{r, include=FALSE}
knitr::opts_chunk$set(echo = T, warning = F, message = F, eval = T, results = 'hide',
                      fig.align = 'center', fig.width = 8, fig.height = 8)
knitr::opts_knit$set(root.dir = here::here())
```

```{r}

source("R/00-functions_basic.R")
source("R/00-functions_maps.R") 
source("R/00-MCMC_diagnostics.R")
library(tidyverse)
library(sf)
library(srvyr) 
library(rgeos)
library(raster)
library(PrevMap)
set.seed(48)


#......................
# Import Data
#......................
dt <- readRDS("data/derived_data/vividepi_recode_completecases.rds")
dtsrvy <- makecd2013survey(survey = dt)
ge <- readRDS(file = "data/raw_data/dhsdata/VivIDge.RDS")
DRCprov <- readRDS("data/map_bases/vivid_DRCprov.rds")

```

```{r}

#...................... 
# get long lat back
#......................
longlat <- dt %>% 
  dplyr::select(c("hv001", "longnum", "latnum")) %>% 
  dplyr::filter(!duplicated(.))
#......................
# Subset to Pv
#......................
pvclust.weighted <- dtsrvy %>% 
  dplyr::mutate(count = 1) %>% 
  dplyr::group_by(hv001) %>% 
  dplyr::summarise(n = srvyr::survey_total(count), 
                   plsmdn = srvyr::survey_total(pv18s, na.rm = T), 
                   plsmdprev = srvyr::survey_mean(pv18s, na.rm = T, vartype = c("se", "ci"), level = 0.95)) %>% 
  dplyr::left_join(., ge,  by = "hv001") %>% 
  dplyr::left_join(., longlat, by = "hv001")

# vectors have destroyed spatial class, need to remake
pvclust.weighted <- sf::st_as_sf(pvclust.weighted)
# sanity
identical(sf::st_crs(pvclust.weighted), sf::st_crs(ge))

# need to keep integers, so will round
pvclust.weighted <- pvclust.weighted %>% 
  dplyr::mutate(plsmdn = round(plsmdn, 0),
                n = round(n, 0))


```

```{r}

#-------------------------------------------------------------------------
# Aggregate Covariates
#-------------------------------------------------------------------------
# precipitation already in dataset 
pvclst.covar <- dt %>% 
  dplyr::select("hv001", "precip_mean_cont_scale_clst") %>% 
  dplyr::filter(!duplicated(.))

# bring in health distance
hlthdist <- readRDS("data/derived_data/vividepi_hlthdist_clstmeans.rds") %>% 
  dplyr::select(c("hv001", "hlthdist_cont_scale_clst"))

pvclst.covar <- dplyr::left_join(x = pvclst.covar, y = hlthdist, by = "hv001")

# join together
pvclust.weighted <- dplyr::left_join(pvclust.weighted, pvclst.covar, by = "hv001")
pvclust.weighted.nosf <- pvclust.weighted
sf::st_geometry(pvclust.weighted.nosf) <- NULL

riskvars = c("precip_mean_cont_scale_clst", "hlthdist_cont_scale_clst")


```

## Spatial Autocorrelation EDA for Gaussian Processes
Note, using the vignette/resource provided by [Giorgi & Diggle 2017](https://www.jstatsoft.org/article/view/v078i08) and `rgeos`.


```{r}
uvec <- c(seq(0, 1, by = 0.2), seq(1, 10))
```

### Case Counts
```{r, results='asis'}

pvclust.weighted.geodata <- geoR::as.geodata(obj = pvclust.weighted.nosf,
                                             coords.col = c("longnum", "latnum"),
                                             data.col = "plsmdn",
                                             covar.col = riskvars)

plot(pvclust.weighted.geodata)

```


#### Case-Variogram

```{r, results='asis'}

pv.vario.counts <- geoR::variog(pvclust.weighted.geodata,
                                bin.cloud = T,
                                uvec = uvec)

par(mfrow = c(1,3))
points(pvclust.weighted.geodata, xlab = "Coord X", 
       ylab = "Coord Y",
       cex.max = 0.7, col = gray(seq(1, 0.1, l = 100)),
       pt.divide = "equal")
plot(pv.vario.counts)

plot(pv.vario.counts, bin.cloud = T)
mtext("Empirical Variogram for P.vivax Counts", side = 3,  line = -6, outer = TRUE)

graphics.off()

```

This does not indicate a great degree (if any) spatial structuring. The classic variogram should follow a log-curve and have a sill that is reached at some distance. Of note, the potential nugget effect here is huge.  


### Prevalence
N.B. I am using a empirical logit transformation here. However, in the Bayesian models, I am using an exact approach for the binomial counts. This is purely so we can model this as a normal process (and speed thnings up/use standard regression approachs for EDA).
```{r, results='asis'}
# transform count of "successes" to logit space
pvclust.weighted.nosf$plsmdlogit <- log( (pvclust.weighted.nosf$plsmdn + 0.5)/(pvclust.weighted.nosf$n - pvclust.weighted.nosf$plsmdn + 0.5) ) # 0.5 as tolerance for 0s

pvclust.weighted.geodata <- geoR::as.geodata(obj = pvclust.weighted.nosf,
                                             coords.col = c("longnum", "latnum"),
                                             data.col = "plsmdlogit",
                                             covar.col = riskvars)
plot(pvclust.weighted.geodata)


```




#### Prevelance-Variogram

```{r, results='asis'}

pv.vario.prev <- geoR::variog(pvclust.weighted.geodata,
                              bin.cloud = T,
                              uvec = uvec)

par(mfrow = c(1,3))
points(pvclust.weighted.geodata, xlab = "Coord X", 
       ylab = "Coord Y",
       cex.max = 0.7, col = gray(seq(1, 0.1, l = 100)),
       pt.divide = "equal")
plot(pv.vario.prev)

plot(pv.vario.prev, bin.cloud = T)

mtext("Empirical Variogram for P.vivax Prevelance", side = 3,  line = -6, outer = TRUE)

graphics.off()

```
A remarkable lack of spatial structure...?


#### Matern Shape
```{r}
#......................
# Matern Shape, K
#......................

# look at kappa smooth parameter for matern covariance
profile.kappa <- invisible( PrevMap::shape.matern(plsmdlogit ~ 1,
                                                  coords = ~ longnum + latnum,
                                                  data = pvclust.weighted.nosf,
                                                  set.kappa = seq(0.1, 10, length = 30),
                                                  start.par = c(0.2, 0.05), # starting values for the scale parameter phi and the relative variance of the nugget effect nu2
                                                  coverage = NULL # CI coverage
) )



```

```{r, results='asis'}
plot(profile.kappa)
abline(v = 0.76, col = "blue", lty = 2)
```
The max _kappa_ value was found to be `r profile.kappa$set.kappa[ which(profile.kappa$val.kappa == max(profile.kappa$val.kappa)) ]`. Will set it to 0.75 (fixed now in subsequent analyses).

#### Estimation of Variogram Parameters
I.e. nugget effect, sill, and range. Will test a Matern and Exponential Covariance and compare. 
```{r}

#..........................
# Maximum Likelihood
#..........................
ml.exp <- geoR::likfit(pvclust.weighted.geodata, 
                       ini = c(1, 0.5),
                       cov.model = "exp",
                       fix.nugget = F, 
                       nugget = 0,
                       fix.kappa = T) # matern of kappa of 1 reduces to exp

ml.matern <- geoR::likfit(pvclust.weighted.geodata, 
                          ini = c(1, 0.5),
                          cov.model = "matern",
                          fix.nugget = F, 
                          nugget = 0,
                          fix.kappa = T,
                          kappa = 0.75)

#..........................
# OLS
#..........................
ols.exp <- geoR::variofit(pv.vario.prev, 
                          ini = c(1, 0.5),
                          cov.model = "exp",
                          fix.nugget = F, 
                          nugget = 0,
                          fix.kappa = F) 

ols.matern <- geoR::variofit(pv.vario.prev, 
                             ini = c(1, 0.5),
                             cov.model = "matern",
                             fix.nugget = F, 
                             nugget = 0,
                             fix.kappa = F)



par(mfrow = c(1, 2))
plot(pv.vario.prev, main = "Maximum Likelihood Estimation")
lines(ml.exp, max.dist = 20, lty = 1, lwd = 2)
lines(ml.matern, max.dist = 20, lty = 2, lwd = 2, col = "red")
legend(2, 4, legend = c("Exp", "Matern"), 
       lty = c(1, 2), col = c("black", "red"), cex = 0.7)

plot(pv.vario.prev, main = "Ordinary Least Square")
lines(ols.exp, max.dist = 20, lty = 1, lwd = 2)
lines(ols.matern, max.dist = 20, lty = 2, lwd = 2, col = "red")
legend(2, 4, legend = c("Exp", "Matern"), 
       lty = c(1, 2), col = c("black", "red"), cex = 0.7)

graphics.off()
```

Truly impressive lack of spatial structure. From models above, an appropriate _kappa_ value fro the Matern covariance matrix appeared to be 0.5 or 1. Let's explore that specifically below.

We can also use the information from the empirical variogram fits for choosing starting values in our Bayesian model. 

##### MLE Exponential to the GRF 
```{r, results='asis'}
ml.exp
```

##### MLE Matern to the GRF 
```{r, results='asis'}
ml.matern
```

##### OLS Exponential Empirical Variogram fit
```{r, results='asis'}
ols.exp
```

##### OLS Matern Empirical Variogram fit
```{r, results='asis'}
ols.matern
```

### MLE Estimates w/ PrevMap
#### Intercept Only Model 

```{r}
prevmap.linear.mleintercept <- PrevMap::linear.model.MLE(
  formula=as.formula("plsmdn~1"), 
  coords=as.formula("~longnum+latnum"),
  data=pvclust.weighted.nosf, 
  start.cov.pars=c(1,1),
  kappa=0.75)
```
Summary call for `prevmap.linear.mleintercept` here (print out ugly).
```{r, results='hide', echo = F}
# results print out ugly, so put here
summary(prevmap.linear.mleintercept)
```

##### Potential Priors for Matern function 
```{r, results='asis'}
ests <- summary(prevmap.linear.mleintercept)$cov.pars
expests <- exp(ests)
ests <- cbind.data.frame(intparam = c("sigma2", "phi", "tau2"), 
                         ests = ests,
                         exp_ests = expests)
ests %>% 
  dplyr::mutate_if(is.numeric, round, 3) %>% 
  DT::datatable(., 
                rownames = F,
                options = list(
                  searching = F,
                  pageLength = 7,
                  dom = 'Bfrtip'))



```

#### Covar Model 
```{r}
prevmap.linear.mlecovar <- PrevMap::linear.model.MLE(
  formula=as.formula("plsmdn~precip_mean_cont_scale_clst+hlthdist_cont_scale_clst"), 
  coords=as.formula("~longnum+latnum"),
  data=pvclust.weighted.nosf, 
  start.cov.pars=c(1.5,1),
  kappa=0.75)
```
```{r, results='asis'}
summary(prevmap.linear.mlecovar)
```

##### Potential Priors for Matern function 
```{r, results='asis'}
ests <- summary(prevmap.linear.mlecovar)$cov.pars
expests <- exp(ests)
ests <- cbind.data.frame(covparam = c("sigma2", "phi", "tau2"), 
                         ests = ests,
                         exp_ests = expests)
ests %>% 
  dplyr::mutate_if(is.numeric, round, 3) %>% 
  DT::datatable(., 
                rownames = F,
                options = list(
                  searching = F,
                  pageLength = 7,
                  dom = 'Bfrtip'))

```



## Hierarchial Bayesian Spatial Models for Cluster Prevalence 

### Model Framework 
NB, here it is not quite the same to "turn off space". Instead, we will let the covariance matrix be considered as a continuous process and let $\sigma^2$ and Matern correlation, $\phi$ vary under the model (we fixed $\kappa$). 

#### Priors & Sampling
(Please open up code for inquiries).



### Diagnostic Chains
Run out four diagnostic chains to check convergence. 
```{r}

mod.framework <- tibble::tibble(name = sub("_run_ret.RDS", "", list.files("analyses/07-spatial_prediction/prevmap_diagn_runs/")),
                                path = list.files("analyses/07-spatial_prediction/prevmap_diagn_runs/",
                                                  full.names = TRUE)) %>%   
  dplyr::mutate(mcmc = purrr::map(path, readRDS)) %>% 
  dplyr::select(-c("path"))



```


#### MCMC Diagnostics Chains
Note, our goal is to have the effective sample sizes (`n.effective`) to be at least 500 for all parameters. In addition, we will visually check for convergence among the four chains. 

```{r, results='hide'}
mod.framework$summresults <- invisible( purrr::map(mod.framework$mcmc, summary, hpd.coverage = 0.95) )
mod.framework$summdiagret <- invisible( purrr::pmap(mod.framework[,c("name", "mcmc")], get_diag_summ_results.Bayes.PrevMap) )


mod.framework$paramtable <- invisible( purrr::map(mod.framework$summdiagret, "summresults") )
mod.framework$traceplots <- invisible( purrr::map(mod.framework$summdiagret, "traceplots") ) 
mod.framework$autocorrplot <- invisible( purrr::map(mod.framework$summdiagret, "autocorrplots") )

```

##### Param Diagnostics for Diagnostic Runs
```{r, results='asis'}

mod.framework %>% 
  dplyr::select(c("name", "paramtable")) %>% 
  tidyr::unnest(cols="paramtable") %>% 
  dplyr::mutate_if(is.numeric, round, 3) %>% 
  DT::datatable(., 
                rownames = F,
                options = list(
                  searching = F,
                  pageLength = 7,
                  dom = 'Bfrtip', 
                  buttons = c('csv')))


```



##### MCMC Trace Plots for Diagnostic Runs
Too many plots again to visualize here, so will send out locally.
```{r, results='hide', fig.show='hide'}
#...................... 
# diagnostic chain non-random effects
#......................
diagchain <- "analyses/07-spatial_prediction/cluster_gaus_diagnostic_chain/"
dir.create(diagchain, recursive = T)

for(i in 1:nrow(mod.framework)){
  fileout <- paste0(diagchain, "/", mod.framework$name[i], "_diagnostic_chains_TRACEPLOTS_", i, ".pdf")
  pdf(fileout)
  invisible(lapply(mod.framework$traceplots[i], print)) 
  dev.off()
}


for(i in 1:nrow(mod.framework)){
  fileout <- paste0(diagchain, "/", mod.framework$name[i], "_diagnostic_chains_AUTOPLOTS_", i, ".pdf")
  pdf(fileout)
  invisible(lapply(mod.framework$autocorrplot[i], print)) 
  dev.off()
}

#...................... 
# diagnostic chain non-random effects
#......................
diag1 <- mod.framework$mcmc[mod.framework$name == "covars1.diagnostic"][[1]]
diag2 <- mod.framework$mcmc[mod.framework$name == "covars2.diagnostic"][[1]]
diag3 <- mod.framework$mcmc[mod.framework$name == "covars3.diagnostic"][[1]]
diag4 <- mod.framework$mcmc[mod.framework$name == "covars4.diagnostic"][[1]]


pdf(paste0(diagchain, "diagnostic_chains_random_effect_chains.pdf"))
par(mfrow=c(2,4))
for (i in 1:ncol(diag1$S)) {
  # diag 1 
  diag1x <- 1:nrow(diag1$S)
  diag1y <- diag1$S[,i]
  plot(diag1x, diag1y, type="l")
  
  # diag 2
  diag2x <- 1:nrow(diag2$S)
  diag2y <- diag2$S[,i]
  plot(diag2x, diag2y, type="l")
  
  # diag 3
  diag3x <- 1:nrow(diag3$S)
  diag3y <- diag3$S[,i]
  plot(diag3x, diag3y, type="l")
  
  # diag 4
  diag4x <- 1:nrow(diag4$S)
  diag4y <- diag4$S[,i]
  plot(diag4x, diag4y, type="l")
}
dev.off()


```

##### Gelman Rubin Diagnostic
```{r}
gr_covars <- covariates_gelman.prevmap(
  diag_chain1 = mod.framework$mcmc[[1]],
  diag_chain2 = mod.framework$mcmc[[2]],
  diag_chain3 = mod.framework$mcmc[[3]],
  diag_chain4 = mod.framework$mcmc[[4]]
)

gr_randomeffects <- random_effect_gelman.prevmap(
  diag_chain1 = mod.framework$mcmc[[1]],
  diag_chain2 = mod.framework$mcmc[[2]],
  diag_chain3 = mod.framework$mcmc[[3]],
  diag_chain4 = mod.framework$mcmc[[4]]
)


```

###### Random Effects Gelman
```{r, results='asis'}
gr_randomeffects %>% 
  DT::datatable(., 
                rownames = F,
                options = list(
                  searching = F,
                  pageLength = 20,
                  dom = 'Bfrtip', 
                  buttons = c('csv')))
```

###### Covars Gelman
```{r, results='asis'}

gr_covars %>% 
  DT::datatable(., 
                rownames = F,
                options = list(
                  searching = F,
                  pageLength = 20,
                  dom = 'Bfrtip', 
                  buttons = c('csv')))

```


## Long Chain
```{r}

long.intercept <- readRDS("analyses/07-spatial_prediction/prevmap_long_runs/intercept_model.RDS")
long.covar <- readRDS("analyses/07-spatial_prediction/prevmap_long_runs/covariate_model.RDS")
long.framework <- tibble::tibble(name = c("intercept", "covars"),
                                 mcmc = list(long.intercept, long.covar))

```

### Final Params
```{r, results='hide', fig.show='hide'}

long.framework$summresults <- purrr::map(long.framework$mcmc, summary, hpd.coverage = 0.95)
long.framework$summdiagret <- invisible( purrr::pmap(long.framework[,c("name", "mcmc")], get_diag_summ_results.Bayes.PrevMap) )
mod.framework

long.framework$paramtable <- purrr::map(long.framework$summdiagret, "summresults")

long.frameworks.summresults <- long.framework %>% 
  dplyr::select(c("name", "paramtable")) %>% 
  tidyr::unnest(cols="paramtable") 
```

```{r, results='asis'}
long.frameworks.summresults %>% 
  dplyr::mutate_if(is.numeric, round, 2) %>% 
  DT::datatable(., 
                rownames = F,
                options = list(
                  searching = F,
                  pageLength = 7,
                  dom = 'Bfrtip', 
                  buttons = c('csv')))


```

### Final Diag Plots
#### Intercept Only Diagnostics 
```{r, results='asis'}
PrevMap::trace.plot(long.intercept, param = "beta", component.beta = 1)
PrevMap::autocor.plot(long.intercept, param = "beta", component.beta = 1)

PrevMap::trace.plot(long.intercept, param = "tau2")
PrevMap::autocor.plot(long.intercept, param = "tau2")

PrevMap::trace.plot(long.intercept, param = "sigma2")
PrevMap::autocor.plot(long.intercept, param = "sigma2")

PrevMap::trace.plot(long.intercept, param = "phi")
PrevMap::autocor.plot(long.intercept, param = "phi")

```

#### Covariate Model Diagnostics 
```{r, results='asis'}
PrevMap::trace.plot(long.covar, param = "beta", component.beta = 1)
PrevMap::autocor.plot(long.covar, param = "beta", component.beta = 1)


PrevMap::trace.plot(long.covar, param = "beta", component.beta = 2)
PrevMap::autocor.plot(long.covar, param = "beta", component.beta = 2)

PrevMap::trace.plot(long.covar, param = "beta", component.beta = 3)
PrevMap::autocor.plot(long.covar, param = "beta", component.beta = 3)


PrevMap::trace.plot(long.covar, param = "tau2")
PrevMap::autocor.plot(long.covar, param = "tau2")

PrevMap::trace.plot(long.covar, param = "sigma2")
PrevMap::autocor.plot(long.covar, param = "sigma2")

PrevMap::trace.plot(long.covar, param = "phi")
PrevMap::autocor.plot(long.covar, param = "phi")
```

##### Random Effects for Covars
Too large to plot, so will send out locally.
```{r}
#............................................................
# send out random effects to local drive
#...........................................................
dir.create("analyses/07-spatial_prediction/cluster_gaus_long_chain",
           recursive = T)
pdf("analyses/07-spatial_prediction/cluster_gaus_long_chain/cluster_gaus_long_chain_chain_random_effect_chains.pdf")
par(mfrow=c(4,1))
for(i in 1:ncol(long.covar$S)) {
  plot(coda::as.mcmc(long.covar$S[,i]))
}
dev.off()

#...................... 
# get min effective N
#......................
eff_n_random_effs <- apply(long.covar$S, 2, function(x){
  coda::effectiveSize(coda::as.mcmc(x))
})

```
The minimum effective N among the covariate cluster random effects is: `r min(eff_n_random_effs)`. 

<!-- ## DIC of Final Models -->
<!-- ```{r} -->

<!-- # note because the model returns the S matrix which is the posterior for the random effect -->
<!-- # and it returns the posterior betas, we can calculate the Ti from each model  -->
<!-- # from there, it is straightforward to calculate the likelihood as a series of  -->
<!-- # independently drawn binomial trials -->

<!-- long.framework$randomeff <- purrr::map(long.framework$mcmc, "S") -->
<!-- long.framework$betas <- purrr::map(long.framework$mcmc, "estimate") -->
<!-- long.framework$betas <- purrr::map(long.framework$betas, function(x){return(x[, colnames(x)[!colnames(x) %in% c("sigma^2", "phi", "tau^2")] ] )}) -->
<!-- long.framework$X <- purrr::map(long.framework$mcmc, "D") # matrix of covariates -->
<!-- long.framework$Yi <- lapply(1:nrow(long.framework), function(x) return(pvclust.weighted$plsmdn)) # binomial successes -->
<!-- long.framework$ni <- lapply(1:nrow(long.framework), function(x) return(pvclust.weighted$n)) # observations  -->

<!-- long.framework$Ti <- purrr::pmap(long.framework[,c("X", "betas", "randomeff")],  -->
<!--                                  function(X, betas, randomeff){ -->
<!--                                    bi <- X %*% t(betas) # get the X (nxp matrix) * the betas (pxsim matrix) -->
<!--                                    Ti <- randomeff + t(bi) # random effects is in sim x n so tranpose bi to match it -->
<!--                                    return(Ti) -->
<!--                                  }) -->

<!-- long.framework$LL <- purrr::pmap(long.framework[,c("Yi", "ni", "Ti")], function(Yi, ni, Ti){ -->
<!--   LL <- c() -->
<!--   for(s in 1:nrow(Ti)){ # for every sim -->
<!--     LL.iter <- 0 -->
<!--     for(i in 1:ncol(Ti)){ # for every spatial unit -->
<!--       LL.iter <- dbinom(x = Yi[i], size = ni[i], prob = expit(Ti[s, i]), log = T) + LL.iter -->
<!--       LL.iter -->
<!--     } -->
<!--     LL <- append(LL.iter, LL) -->

<!--   } -->

<!--   LL <- return(LL) -->
<!-- }) -->


<!-- long.framework$DICg <- purrr::map(long.framework$LL, function(x){ -->
<!--   mu = mean(x) -->
<!--   sigma = var(x)/2 -->
<!--   DICg = mu + sigma -->
<!--   return(DICg) -->
<!-- }) -->


<!-- ``` -->

<!-- ```{r, results = 'asis'} -->

<!-- long.framework %>%  -->
<!--   dplyr::select(c("name", "DICg")) %>%  -->
<!--   tidyr::unnest(cols="DICg")  %>%  -->
<!--   dplyr::mutate(DICgchar = as.character(DICg)) %>%  # odd rounding behavior -->
<!--   dplyr::mutate(DICg = round(DICg, 2)) %>%  -->
<!--   magrittr::set_colnames(c("Model", "DICg", "DICgchar")) %>%  -->
<!--   kableExtra::kable(format = "html", escape = F) %>% -->
<!--   kableExtra::kable_styling("striped", full_width = F)  -->


<!-- ``` -->

<!-- ### Log Likelihood Trace Plots for Final Models -->
<!-- ```{r, results = 'asis'} -->

<!-- plot(long.framework$LL[[1]], main = "Intercept Model", type = "l") -->
<!-- plot(long.framework$LL[[2]], main = "Covars Model", type = "l") -->

<!-- ``` -->
