==> analyses/07-spatial_prediction/03-Pv_get_predictions_frontend.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

# Bayesian Hierarchial Spatial Model Predictions 

```{r, include=FALSE}
knitr::opts_chunk$set(echo = T, warning = F, message = F, eval = T, results = 'hide',
                      fig.align = 'center')
knitr::opts_knit$set(root.dir = here::here())
```

```{r}
library(tidyverse)
library(CARBayes)
library(PrevMap)
library(ggspatial)
set.seed(48)

load("data/map_bases/vivid_maps_bases.rda")
drccites <- readr::read_csv("data/map_bases/DRC_city_coordinates.csv") %>% 
  dplyr::filter(population > 1e6 | city %in% c("Kisangani"))
drccites <- drccites %>% 
  dplyr::mutate(city = stringr::str_split_fixed(city, pattern = " ", n = 2)[,1])

```


## Spatial Predictions
Fitted values are taken from the Hierarchial Bayesian Spatial Models that were previously fit and determined to be the best model choice by model fit (Deviance Information Criteria).

### Province Models
```{r}
# import data
DRCprov <- readRDS("data/map_bases/vivid_DRCprov.rds")
provmodel <- readRDS("analyses/07-spatial_prediction/ProvModels/ProvModel_long_chains.RDS")
provmodel <- provmodel[provmodel$name == "CAR_covar", ] # a priori model

```

#### Fitted Values (Posterior Means)
```{r}
provs <- provmodel$data[[1]]$adm1name
provmodel.samples.fittedvalues <- provmodel$MCMC[[1]]$samples$fitted
provmodel.posteriormeans <- apply(provmodel.samples.fittedvalues, 2, mean)
```

```{r, results='asis'}
provmodel.posteriormeans <- cbind.data.frame(adm1name = provs, posteriormeanscounts = provmodel.posteriormeans) 
rownames(provmodel.posteriormeans) <- NULL

provmodel.posteriormeans %>% 
  dplyr::mutate_if(., is.numeric, round, 4) %>% 
  DT::datatable(., extensions='Buttons',
                options = list(
                  searching = T,
                  pageLength = 10,
                  dom = 'Bfrtip', 
                  buttons = c('csv')))
```


```{r, results='asis'}
provmodel.posteriormeans <- cbind.data.frame(adm1name = provs, posteriormeansprev = provmodel.posteriormeans$posteriormeanscounts) 
provmodel.posteriormeans$posteriormeansprev <- provmodel.posteriormeans$posteriormeansprev/provmodel$data[[1]]$n
rownames(provmodel.posteriormeans) <- NULL

provmodel.posteriormeans %>% 
  dplyr::mutate_if(is.numeric, round, 4) %>% 
  DT::datatable(., extensions='Buttons',
                options = list(
                  searching = T,
                  pageLength = 10,
                  dom = 'Bfrtip', 
                  buttons = c('csv')))
```


```{r}
#........................
# Bring it home
#........................
# adm1names
DRCprovcentroids <- tibble::tibble(ad1name = DRCprov$adm1name,
                                   cntr = sf::st_centroid(DRCprov$geometry)) %>% 
  dplyr::mutate(longnum = sf::st_coordinates(cntr)[,1],
                latnum = sf::st_coordinates(cntr)[,2])
# plot df
provmodelplotObjdf <- dplyr::left_join(DRCprov, provmodel.posteriormeans) 
provmodelplotObj <- ggplot() + 
  geom_sf(data = provmodelplotObjdf, 
          aes(fill = posteriormeansprev), color = "#737373") +
  scale_fill_distiller("Prevalence", type = "div", palette = "RdYlBu",
                       na.value = NA) +
  coord_sf(datum=NA) +  # to get rid of gridlines
  prettybasemap_nodrc_dark_nonorth +
  # ggrepel::geom_text_repel(data = DRCprovcentroids,
  #                          aes(x = longnum,  y = latnum, label = ad1name),
  #                          fontface = "bold", alpha = 0.5) +
  theme(legend.position = "right",
        legend.title = element_text(size = 13, face = "bold", angle = 0, vjust = 0.5, hjust = 0.5),
        legend.text = element_text(size = 12, face = "bold", angle = 0))

jpeg(file = "results/figures/ProvMap_postMeans.jpg",
     height = 11, width = 8, units = "in", res = 500)
plot(provmodelplotObj)
graphics.off()

# save out for final map
saveRDS(provmodelplotObj, "results/ProvMap_postMeans_PlotObj.RDS")

```

```{r, results='asis'}
plot(provmodelplotObj)
```

#### Standard Errors
```{r}
# following assumption diggle and giorgi make here: https://github.com/cran/PrevMap/blob/master/R/foo.R
provmodel.samples.fittedvalues_prevs <- apply(provmodel.samples.fittedvalues, 
                                              1, function(x){x/provmodel$data[[1]]$n})
provmodel.se <- apply(provmodel.samples.fittedvalues_prevs, 1, sd)

```

```{r, results='asis'}
provmodel.posteriorse <- cbind.data.frame(adm1name = provs, posteriorses = provmodel.se) 
rownames(provmodel.posteriorse) <- NULL

provmodel.posteriorse %>% 
  dplyr::mutate_if(is.numeric, round, 4) %>% 
  DT::datatable(., extensions='Buttons',
                options = list(
                  searching = T,
                  pageLength = 10,
                  dom = 'Bfrtip', 
                  buttons = c('csv')))
```

```{r}
provmodel.posteriorseplotdf <- dplyr::left_join(DRCprov, provmodel.posteriorse) 


provmodel.posteriorselotObj <- ggplot() +
  geom_sf(data = provmodel.posteriorseplotdf, 
          aes(fill = posteriorses), color = "#737373") +
  scale_fill_distiller("Standard Error", type = "div", palette = "RdYlBu",
                       na.value = NA) +
  coord_sf(datum=NA) +
  theme_void() +
  theme(legend.position = "right",
        legend.title = element_text(size = 13, face = "bold", angle = 0, vjust = 0.5, hjust = 0.5),
        legend.text = element_text(size = 12, face = "bold", angle = 0))

jpeg(file = "results/figures/ProvMap_postSEs.jpg",
     height = 11, width = 8, units = "in", res = 500)
plot(provmodel.posteriorselotObj)
graphics.off()

# save out for final map
saveRDS(provmodel.posteriorselotObj, "results/ProvMap_postSE_PlotObj.RDS")

```
```{r, results='asis'}
plot(provmodel.posteriorselotObj)
```


### Cluster Models 

#### Fitted Values (Posterior Means)

```{r}
# A "pred.PrevMap" object list with the following components: logit; prevalence; odds; probit;exceedance.prob, corresponding to a matrix of the exceedance probabilities where each column corresponds to a specified value in thresholds; samples, corresponding to a matrix of the posterior samples at each prediction locations for the linear predictor; grid.pred prediction locations. Each of the three components logit, prevalence, odds and probit is also a list with the following components:
source("R/00-functions_basic.R")
# for mask sanity
DRC <- readRDS("data/map_bases/gadm/gadm36_COD_0_sp.rds")

# read rds
clstmodel <- readRDS("analyses/07-spatial_prediction/prevmap_predictions/covar_predictions.RDS")
clstmodel.fitted <- expit(clstmodel$samples)
clstmodel.coords <- clstmodel$grid.pred
clstmodel.fitted.postmeans <- apply(clstmodel.fitted, 2, mean)
clstmodel.coords.means <- cbind.data.frame(clstmodel.coords, 
                                           fitted.postmean = clstmodel.fitted.postmeans)

#...................... 
# sanity check
#......................
sanitycheck <- clstmodel$prevalence$predictions
if ( !all.equal(sanitycheck, clstmodel.coords.means$fitted.postmean) ) {
  stop("Issue with prevmap understanding")
}

#..............................................................
# Make Raster from PrevMap 
#..............................................................
clstmodel.fitted.postmeans.raster <- 
  raster::rasterFromXYZ(cbind(clstmodel.coords.means[, "longnum"],
                              clstmodel.coords.means[, "latnum"],
                              clstmodel.coords.means[, "fitted.postmean"]),
                        crs="+proj=longlat +datum=WGS84 +no_defs")
# crop out values
sp::identicalCRS(clstmodel.fitted.postmeans.raster, sf::as_Spatial(DRC))
sanity_clstmodel.fitted.postmeans.raster <- raster::mask(clstmodel.fitted.postmeans.raster, DRC)
sanity_clstmodel.fitted.postmeans <- values(clstmodel.fitted.postmeans.raster)
summary(sanity_clstmodel.fitted.postmeans)
summary(clstmodel.fitted.postmeans)
# good
```

```{r, results='asis'}

clstmodel.fitted.postmeansdf <- data.frame(PrevPred =clstmodel.fitted.postmeans[!is.na(clstmodel.fitted.postmeans)])
clstmodel.fitted.postmeansdf %>% 
  dplyr::mutate_if(is.numeric, round, 4) %>% 
  DT::datatable(., extensions='Buttons',
                options = list(
                  searching = T,
                  pageLength = 10,
                  dom = 'Bfrtip', 
                  buttons = c('csv')))
```

```{r}

ret.smrstr.m.plot <- ggplot() + 
  ggspatial::layer_spatial(data = clstmodel.fitted.postmeans.raster, aes(fill = stat(band1))) +
  scale_fill_distiller("Prevalence", type = "div", palette = "RdYlBu",
                       na.value = NA) +
  prettybasemap_nodrc_dark_nonorth +
  geom_point(data = drccites, aes(x = longnum, y=latnum), alpha = 0.5) + 
  geom_text(data = drccites, aes(label = city, x = longnum, y=latnum), 
            hjust = 0.5, vjust = 0.5, nudge_y = 0.5, fontface = "bold", alpha = 0.5) +
  theme(legend.position = "right",
        legend.title = element_text(size = 13, face = "bold", angle = 0, vjust = 0.5, hjust = 0.5),
        legend.text = element_text(size = 12, face = "bold", angle = 0))

# plot
jpeg(file = "results/figures/ClstMap_postMeans_interpolate.jpg",
     height = 11, width = 8, units = "in", res = 500)
plot(ret.smrstr.m.plot)
graphics.off()

# save out for final map
saveRDS(ret.smrstr.m.plot, "results/ClustPrevMap_postMeans_PlotObj.RDS")

```

```{r, results='asis'}
plot(ret.smrstr.m.plot)
```


```{r, results='asis', fig.align = 'center', fig.width = 11, fig.height = 8}
#..............................................................
# figure together
#..............................................................
cowplot::plot_grid(provmodelplotObj, ret.smrstr.m.plot,
                   align = "h", nrow = 1)

jpeg(file = "results/figures/Prov_and_ClstMap_postMeans_interpolate.jpg",
     height = 8, width = 11, units = "in", res = 500)
cowplot::plot_grid(provmodelplotObj, ret.smrstr.m.plot,
                   align = "h", nrow = 1, labels = c("(A)", "(B)"))
graphics.off()


svglite::svglite(file = "results/figures/Prov_and_ClstMap_postMeans_interpolate.svg",
                 height = 8, width = 11)
cowplot::plot_grid(provmodelplotObj, ret.smrstr.m.plot,
                   align = "h", nrow = 1, labels = c("(A)", "(B)"))
graphics.off()


```


#### Standard Errors

```{r}
# borrowed from plot.pred.Prevmap overload on github: https://github.com/cran/PrevMap/blob/master/R/foo.R
clstmodel.fitted.se <- rasterFromXYZ(cbind(clstmodel$grid,
                                           clstmodel[[1]][["standard.errors"]]))

```

```{r, results='asis'}

clstmodel.fitted.sedf <- data.frame(PrevSE = values(clstmodel.fitted.se))
minse <- min(clstmodel.fitted.sedf$PrevSE, na.rm = T)
maxse <- max(clstmodel.fitted.sedf$PrevSE, na.rm = T)

knitr::kable(tibble::tibble(nms = c("minSE", "maxSE"), ret = c(minse, maxse)))

```


```{r}
clstmodel.fitted.se.raster <- clstmodel.fitted.se
crs(clstmodel.fitted.se.raster) <- "+init=epsg:4326"

ret.smrstr.m.plot <- ggplot() + 
  ggspatial::layer_spatial(data = clstmodel.fitted.se.raster, aes(fill = stat(band1)),
                           interpolate = TRUE) +
  scale_fill_distiller("Standard Error", type = "div", palette = "RdYlBu", 
                       na.value = NA) +
  theme_void() +
  theme(legend.position = "right",
        legend.title = element_text(size = 13, face = "bold", angle = 0, vjust = 0.5, hjust = 0.5),
        legend.text = element_text(size = 12, face = "bold", angle = 0))


# plot
jpeg(file = "results/figures/ClstPrevmapMap_postSE_interpolate.jpg",
     height = 11, width = 8, units = "in", res = 500)
plot(ret.smrstr.m.plot)
graphics.off()
# save out for final map
saveRDS(ret.smrstr.m.plot, "results/ClustPrevMap_postSE_PlotObj.RDS")


```

```{r, results='asis'}
plot(ret.smrstr.m.plot)
```



```{r, results='asis', fig.align = 'center', fig.width = 11, fig.height = 8}
#..............................................................
# figure for paper
#..............................................................
cowplot::plot_grid(provmodel.posteriorselotObj, ret.smrstr.m.plot,
                   align = "h", nrow = 1)

jpeg(file = "results/figures/Prov_and_ClstMap_postSE.jpg",
     height = 8, width = 11, units = "in", res = 500)
cowplot::plot_grid(provmodel.posteriorselotObj, ret.smrstr.m.plot,
                   align = "h", nrow = 1, labels = c("(A)", "(B)"))
graphics.off()


svglite::svglite(file = "results/figures/Prov_and_ClstMap_postSE.svg",
                 height = 8, width = 11)
cowplot::plot_grid(provmodel.posteriorselotObj, ret.smrstr.m.plot,
                   align = "h", nrow = 1, labels = c("(A)", "(B)"))
graphics.off()



```

### Cluster Model Covariates
```{r}
covar.rstr.pred <- readRDS("data/derived_data/vividepi_spatial_covar_feature_engineer.rds")

precip.plot <- ggplot() + 
  ggspatial::layer_spatial(data = covar.rstr.pred[["precip_mean_cont_scale_clst"]], 
                                        aes(fill = stat(band1))) +
  scale_fill_viridis_c("Precip", na.value = NA) +
  theme_void()


hlthdist.plot <- ggplot() + 
  ggspatial::layer_spatial(data = covar.rstr.pred[["hlthdist_cont_scale_clst"]], 
                                          aes(fill = stat(band1))) +
  scale_fill_viridis_c("Health Dist.", na.value = NA) +
  theme_void()

#......................
# figure for paper
#......................
jpeg(file = "results/figures/Clst_Covariates.jpg",
     height = 11, width = 8, units = "in", res = 800)
cowplot::plot_grid(precip.plot, hlthdist.plot,
                   align = "h", nrow = 1)
graphics.off()

```
```{r, results='asis'}
cowplot::plot_grid(precip.plot, hlthdist.plot,
                   align = "h", nrow = 1)
```



## Appendix 
### Same Scale as Pf
```{r}
clstmodel.fitted.postmeans.raster <- 
  raster::rasterFromXYZ(cbind(clstmodel.coords[, "longnum"],
                              clstmodel.coords[, "latnum"],
                              clstmodel.fitted.postmeans),
                        crs="+proj=longlat +datum=WGS84 +no_defs")



ret.smrstr.m.plot <- ggplot() + 
  ggspatial::layer_spatial(data = clstmodel.fitted.postmeans.raster, aes(fill = stat(band1))) +
  scale_fill_distiller("Prevalence", type = "div", palette = "RdYlBu", 
                       na.value = NA,
                       limits = c(0, 0.6)) +
  smpl_base_map +
  geom_point(data = drccites, aes(x = longnum, y=latnum)) + 
  geom_text(data = drccites, aes(label = city, x = longnum, y=latnum), 
            hjust = 0.5, vjust = 0.5, nudge_y = 0.5, fontface = "bold") +
  theme(legend.position = "right",
        legend.title = element_text(size = 13, face = "bold", angle = 0, vjust = 0.5, hjust = 0.5),
        legend.text = element_text(size = 12, face = "bold", angle = 0))


jpeg(file = "results/figures/ClstMap_postMeans_samescale_pf.jpg",
     height = 11, width = 8, units = "in", res = 500)
plot(ret.smrstr.m.plot)
graphics.off()

```

```{r, results='asis'}
plot(ret.smrstr.m.plot)
```


### Reasonable Scale 
```{r}
clstmodel.fitted.postmeans.raster <- 
  raster::rasterFromXYZ(cbind(clstmodel.coords[, "longnum"],
                              clstmodel.coords[, "latnum"],
                              clstmodel.fitted.postmeans),
                        crs="+proj=longlat +datum=WGS84 +no_defs")



ret.smrstr.m.plot <- ggplot() + 
  ggspatial::layer_spatial(data = clstmodel.fitted.postmeans.raster, aes(fill = stat(band1))) +
  scale_fill_distiller("Prevalence", type = "div", palette = "RdYlBu", 
                       na.value = NA,
                       limits = c(0, 0.05)) +
  smpl_base_map +
  theme(legend.position = "right",
        legend.title = element_text(size = 13, face = "bold", angle = 0, vjust = 0.5, hjust = 0.5),
        legend.text = element_text(size = 12, face = "bold", angle = 0))


jpeg(file = "results/figures/ClstMap_postMeans_reasonable_scale.jpg",
     height = 11, width = 8, units = "in", res = 500)
plot(ret.smrstr.m.plot)
graphics.off()

```

```{r, results='asis'}
plot(ret.smrstr.m.plot)
```


#### With Nonzero Cluster Prev Points
This excludes prevalences of 0
```{r}
mp <- readRDS( "data/derived_data/basic_cluster_mapping_data.rds")
pvclst <- mp[mp$plsmdmspec == "pv18s" & mp$maplvl == "hv001", ]
pvclst <- pvclst$data[[1]] %>% 
  dplyr::filter(plsmdprev > 0)

clst.points <- ggplot() + 
  ggspatial::layer_spatial(data = clstmodel.fitted.postmeans.raster, aes(fill = stat(band1))) +
  scale_fill_distiller("Prevalence", type = "div", palette = "RdYlBu",
                       na.value = NA) +
  smpl_base_map +
  geom_point(data = pvclst, aes(x=longnum, y=latnum, color = plsmdprev), 
             alpha = 0.5, shape = 16) + 
  scale_color_viridis_c() + 
  theme(legend.position = "right",
        legend.title = element_text(size = 13, face = "bold", angle = 0, vjust = 0.5, hjust = 0.5),
        legend.text = element_text(size = 12, face = "bold", angle = 0))


jpeg(file = "results/figures/ClstMap_postMeans_interpolate_with_clst_onlyposprev.jpg",
     height = 11, width = 8, units = "in", res = 500)
plot(clst.points)
graphics.off()


```
```{r, results='asis'}
plot(clst.points)
```

#### With Nonzero Cluster Prev Points and no outliers
This excludes prevalences of 0 and prevalences over 0.2 for a more reasonable scale
```{r}
mp <- readRDS( "data/derived_data/basic_cluster_mapping_data.rds")
pvclst <- mp[mp$plsmdmspec == "pv18s" & mp$maplvl == "hv001", ]
pvclst <- pvclst$data[[1]] %>% 
  dplyr::filter(plsmdprev > 0) %>% 
  dplyr::filter(plsmdprev < 0.2)

clst.points <- ggplot() + 
  ggspatial::layer_spatial(data = clstmodel.fitted.postmeans.raster, aes(fill = stat(band1))) +
  scale_fill_distiller("Prevalence", type = "div", palette = "RdYlBu",
                       na.value = NA) +
  smpl_base_map +
  geom_point(data = pvclst, aes(x=longnum, y=latnum, color = plsmdprev), 
             alpha = 0.5, shape = 16) + 
  scale_color_viridis_c() + 
  theme(legend.position = "right",
        legend.title = element_text(size = 13, face = "bold", angle = 0, vjust = 0.5, hjust = 0.5),
        legend.text = element_text(size = 12, face = "bold", angle = 0))


jpeg(file = "results/figures/ClstMap_postMeans_interpolate_with_clst_onlyposprev.jpg",
     height = 11, width = 8, units = "in", res = 500)
plot(clst.points)
graphics.off()


```
```{r, results='asis'}
plot(clst.points)
```



### Intercept Comparison
#### Province

```{r}
provmodel <- readRDS("analyses/07-spatial_prediction/ProvModels/ProvModel_long_chains.RDS")
provmodel <- provmodel[provmodel$name == "CAR_intercept", ] # compare model
provs <- provmodel$data[[1]]$adm1name
provmodel.samples.fittedvalues <- provmodel$MCMC[[1]]$samples$fitted
provmodel.posteriormeans <- apply(provmodel.samples.fittedvalues, 2, mean)
provmodel.posteriormeans <- cbind.data.frame(adm1name = provs, posteriormeanscounts = provmodel.posteriormeans) 
provmodel.posteriormeans <- cbind.data.frame(adm1name = provs, posteriormeansprev = provmodel.posteriormeans$posteriormeanscounts) 
provmodel.posteriormeans$posteriormeansprev <- provmodel.posteriormeans$posteriormeansprev/provmodel$data[[1]]$n

# plot df
provmodelplotObjdf <- dplyr::left_join(DRCprov, provmodel.posteriormeans) 
provmodelplotObj <- ggplot() + 
  geom_sf(data = provmodelplotObjdf, 
          aes(fill = posteriormeansprev), color = "#737373") +
  scale_fill_distiller("Prevalence", type = "div", palette = "RdYlBu",
                       na.value = NA) +
  coord_sf(datum=NA) +  # to get rid of gridlines
  prettybasemap_nodrc_dark_nonorth +
  theme(legend.position = "right",
        legend.title = element_text(size = 13, face = "bold", angle = 0, vjust = 0.5, hjust = 0.5),
        legend.text = element_text(size = 12, face = "bold", angle = 0))
```

```{r, results='asis'}
plot(provmodelplotObj)
```


#### Cluster
```{r}
# read rds
clstmodel <- readRDS("analyses/07-spatial_prediction/prevmap_predictions/intercept_predictions.RDS")
clstmodel.fitted <- expit(clstmodel$samples)
clstmodel.coords <- clstmodel$grid.pred
clstmodel.fitted.postmeans <- apply(clstmodel.fitted, 2, mean)
clstmodel.coords.means <- cbind.data.frame(clstmodel.coords, 
                                           fitted.postmean = clstmodel.fitted.postmeans)
# Make Raster from PrevMap 
clstmodel.fitted.postmeans.raster <- 
  raster::rasterFromXYZ(cbind(clstmodel.coords.means[, "longnum"],
                              clstmodel.coords.means[, "latnum"],
                              clstmodel.coords.means[, "fitted.postmean"]),
                        crs="+proj=longlat +datum=WGS84 +no_defs")

clstmodel.fitted.postmeansdf <- data.frame(PrevPred = clstmodel.fitted.postmeans)
ret.smrstr.m.plot <- ggplot() + 
  ggspatial::layer_spatial(data = clstmodel.fitted.postmeans.raster, aes(fill = stat(band1))) +
  scale_fill_distiller("Prevalence", type = "div", palette = "RdYlBu",
                       na.value = NA) +
  prettybasemap_nodrc_dark_nonorth +
  geom_point(data = drccites, aes(x = longnum, y=latnum), alpha = 0.5) + 
  geom_text(data = drccites, aes(label = city, x = longnum, y=latnum), 
            hjust = 0.5, vjust = 0.5, nudge_y = 0.5, fontface = "bold", alpha = 0.5) +
  theme(legend.position = "right",
        legend.title = element_text(size = 13, face = "bold", angle = 0, vjust = 0.5, hjust = 0.5),
        legend.text = element_text(size = 12, face = "bold", angle = 0))
```
```{r, results='asis'}
plot(ret.smrstr.m.plot)
```
